+++ CHANGELOG.md
@@ -130,6 +130,8 @@
 - Fix checking status when using a `KonnectGatewayControlPlane` with KIC CP type
   as a `ControlPlaneRef`.
   [#1115](https://github.com/Kong/gateway-operator/pull/1115)
+- Fix setting `DataPlane`'s readiness probe using `GatewayConfiguration`.
+  [#1118](https://github.com/Kong/gateway-operator/pull/1118)
 
 ### Added